+++ src/apps/accounts/urls.py
@@ -2,11 +2,12 @@
 
 from django.urls import path
 
-from apps.accounts.views import RegisterView, VerifyOtpView
+from apps.accounts.views import LoginView, RegisterView, VerifyOtpView
 
 app_name = "accounts"
 
 urlpatterns = [
     path("register/", RegisterView.as_view(), name="register"),
     path("verify-otp/", VerifyOtpView.as_view(), name="verify-otp"),
+    path("login/", LoginView.as_view(), name="login"),
 ]